+++ .gitignore
@@ -13,4 +13,3 @@ Release/*
 .vs/*
 build/*
 .vscode/*
-CMakeSettings.json
