+++ CMakeLists.txt
@@ -5,13 +5,14 @@ enable_testing()
 set(CMAKE_COMPILE_WARNING_AS_ERROR ON)
 if (MSVC)
     # needs at least cmake 3.15
-    add_compile_options(/W2 /WX /wd4100)
+    add_compile_options(/W4 /WX)
 else()
     add_compile_options(-Wall -Wextra -Wpedantic -Wno-gnu-empty-struct -Wno-unused-parameter)
 endif()
 
+add_subdirectory(LibraryC)
 add_subdirectory(LibraryCPP)
 add_subdirectory(LibraryCPPClass)
 add_subdirectory(LibraryCPPTemplate)
 
-add_subdirectory(Lab2CPPTemplate)
+add_subdirectory(Lab1C)
