+++ LABS.md
@@ -34,7 +34,7 @@
     add_library(LibraryC STATIC array.c list.c queue.c stack.c vector.c)
     add_subdirectory(Tests)
 
-Первая строчка обисывает сборку подключаемой библиотеки LibraryC из соответствующих исходных файлов.
+Первая строчка описывает сборку подключаемой библиотеки LibraryC из соответствующих исходных файлов.
 Вторая строчка добавляет проекты из подкаталога Tests.
 
 В подкаталоге Tests описаны тесты для базовой проверки работоспособности библиотеки:
