+++ Lab1C/input.txt
@@ -1,4 +1,4 @@
 10
 1 2 3 4 5 6 7 8 9 0
-6
-3 2 5 3 4 1
+4
+1 2 3 4
