+++ Lab1C/lab1.c
@@ -1,0 +1,40 @@
+#include <stdio.h>
+#include "array.h"
+
+Array *array_create_and_read(FILE *input)
+{
+    int n;
+    fscanf(input, "%d", &n);
+    /* Create array */
+    Array *arr = array_create(n, NULL);
+    /* Read array data */
+    for (int i = 0 ; i < n ; ++i)
+    {
+        int x;
+        fscanf(input, "%d", &x);
+        array_set(arr, i, x);
+    }
+    return arr;
+}
+
+void task1(Array *arr)
+{
+}
+
+void task2(Array *arr)
+{
+}
+
+int main(int argc, char **argv)
+{
+    Array *arr = NULL;
+    FILE *input = fopen(argv[1], "r");
+    arr = array_create_and_read(input);
+    task1(arr);
+    array_delete(arr);
+    /* Create another array here */
+    arr = array_create_and_read(input);
+    task2(arr);
+    array_delete(arr);
+    fclose(input);
+}
