--- Lab1CPP/CMakeLists.txt
@@ -1,5 +1,0 @@
-add_executable(Lab1CPP lab1.cpp)
-target_include_directories(Lab1CPP PUBLIC ../LibraryCPP)
-target_link_libraries(Lab1CPP LibraryCPP)
-
-add_test(NAME TestLab1CPP COMMAND Lab1CPP ${CMAKE_CURRENT_SOURCE_DIR}/input.txt)
