+++ LibraryCPP/CMakeLists.txt
@@ -1,3 +1,3 @@
-#add_library(LibraryCPP STATIC array.cpp)
+add_library(LibraryCPP STATIC array.cpp list.cpp queue.cpp stack.cpp vector.cpp)
 
-#add_subdirectory(Tests)
+add_subdirectory(Tests)
