+++ LibraryCPP/Tests/CMakeLists.txt
@@ -1,7 +1,7 @@
-add_executable(TestArrayCPP array.cpp)
-target_include_directories(TestArrayCPP PUBLIC ..)
-target_link_libraries(TestArrayCPP LibraryCPP)
-add_test(TestArrayCPP TestArrayCPP)
+# add_executable(TestArrayCPP array.cpp)
+# target_include_directories(TestArrayCPP PUBLIC ..)
+# target_link_libraries(TestArrayCPP LibraryCPP)
+# add_test(TestArrayCPP TestArrayCPP)
 
 # add_executable(TestListCPP list.cpp)
 # target_include_directories(TestListCPP PUBLIC ..)
