+++ LibraryCPP/stack.h
@@ -17,7 +17,7 @@ void stack_delete(Stack *stack);
 // Should be O(1) on average
 void stack_push(Stack *stack, Data data);
 
-// Retrives the last element from the stack
+// Retrieves the last element from the stack
 Data stack_get(const Stack *stack);
 
 // Removes the last element from the stack
