+++ LibraryCPPClass/stack.h
@@ -24,7 +24,7 @@ public:
     // Should be O(1) on average
     void push(Data data);
 
-    // Retrives the last element from the stack
+    // Retrieves the last element from the stack
     Data get() const;
 
     // Removes the last element from the stack
