+++ LibraryCPPTemplate/Tests/CMakeLists.txt
@@ -11,12 +11,11 @@
 # add_test(TestQueueCPPTemplate TestQueueCPPTemplate)
 # set_tests_properties(TestQueueCPPTemplate PROPERTIES TIMEOUT 10)
 
-add_executable(TestStackCPPTemplate stack.cpp)
-target_include_directories(TestStackCPPTemplate PUBLIC ..)
-add_test(TestStackCPPTemplate TestStackCPPTemplate)
+# add_executable(TestStackCPPTemplate stack.cpp)
+# target_include_directories(TestStackCPPTemplate PUBLIC ..)
+# add_test(TestStackCPPTemplate TestStackCPPTemplate)
 
-add_executable(TestVectorCPPTemplate vector.cpp)
-target_include_directories(TestVectorCPPTemplate PUBLIC ..)
-target_link_libraries(TestVectorCPPTemplate LibraryCPPTemplate)
-add_test(TestVectorCPPTemplate TestVectorCPPTemplate)
-set_tests_properties(TestVectorCPPTemplate PROPERTIES TIMEOUT 10)
+# add_executable(TestVectorCPPTemplate vector.cpp)
+# target_include_directories(TestVectorCPPTemplate PUBLIC ..)
+# add_test(TestVectorCPPTemplate TestVectorCPPTemplate)
+# set_tests_properties(TestVectorCPPTemplate PROPERTIES TIMEOUT 10)
